+++ src/modules/DOM/displayElement/displayTodo.js
@@ -7,47 +7,46 @@ function displayTodo(projectElement, projectData) {
     const todoListData = projectData.todoList;
     todoListData.forEach(todo => {
         const todoSection = document.createElement("section");
+        todoSection.classList.add(`${todo.priority}-priority`);
         const priority = todo.priority;
         todoSection.classList.add(priority);
 
         const listSection = document.createElement("ul");
-        const taskNameItem = document.createElement("li");
 
+        const taskNameAndDueDateItem = document.createElement("li");
         const radioInput = document.createElement("input");
         radioInput.type = "radio";
         radioInput.addEventListener("click", () => {            
             deleteTodo(projectList, projectData.id, todo.id);
             refreshContentDiv();
+            alert("Congratulations");
         })
-
-        const taskNameSpan = document.createElement("span");
-        taskNameSpan.textContent = todo.title;
-        taskNameItem.appendChild(radioInput);
-        taskNameItem.appendChild(taskNameSpan);
+        const taskNameAndDueDateSpan = document.createElement("span");
+        taskNameAndDueDateSpan.textContent = `${todo.title} - ${todo.dueDate}`;
+        taskNameAndDueDateItem.appendChild(radioInput);
+        taskNameAndDueDateItem.appendChild(taskNameAndDueDateSpan);
 
         const descriptionItem = document.createElement("li");
         descriptionItem.textContent = todo.description;
 
-        const dueDateItem = document.createElement("li");
-        dueDateItem.textContent = todo.dueDate;
+        const editAndDeleteItem = document.createElement("li");
+        createEditTodoBtn(editAndDeleteItem,projectData.id, todo);
 
-        const editItem = document.createElement("li");
-        createEditTodoBtn(editItem,projectData.id, todo);
-
-        const deleteItem = document.createElement("li");
         const deleteBtn = document.createElement("button");
         deleteBtn.textContent = "Delete";
         deleteBtn.addEventListener("click", () => {
-            deleteTodo(projectList, projectData.id, todo.id);
-            refreshContentDiv();
+            const result = confirm("Do you want to delete this todo ?");
+            if(result) {
+                deleteTodo(projectList, projectData.id, todo.id);
+                refreshContentDiv();
+                alert("Delete complete");
+            }
         })
-        deleteItem.appendChild(deleteBtn);
+        editAndDeleteItem.appendChild(deleteBtn);
 
-        listSection.appendChild(taskNameItem);
+        listSection.appendChild(taskNameAndDueDateItem);
         listSection.appendChild(descriptionItem);
-        listSection.appendChild(dueDateItem);
-        listSection.appendChild(editItem);
-        listSection.appendChild(deleteItem);
+        listSection.appendChild(editAndDeleteItem);
         
         todoSection.appendChild(listSection);
         projectElement.appendChild(todoSection);
